+++ framework/domain/src/main/java/is/codion/framework/domain/entity/package-info.java
@@ -24,7 +24,7 @@
  * <li>{@link is.codion.framework.domain.entity.Entities#STRICT_DESERIALIZATION}
  * <li>{@link is.codion.framework.domain.entity.attribute.AttributeDefinition#FRACTION_DIGITS}<br>
  * <li>{@link is.codion.framework.domain.entity.attribute.AttributeDefinition#TIME_FORMAT}<br>
- * <li>{@link is.codion.framework.domain.entity.attribute.AttributeDefinition#NUMBER_FORMAT_GROUPING}<br>
+ * <li>{@link is.codion.framework.domain.entity.attribute.AttributeDefinition#NUMBER_GROUPING}<br>
  * <li>{@link is.codion.framework.domain.entity.attribute.AttributeDefinition#GROUPING_SEPARATOR}<br>
  * <li>{@link is.codion.framework.domain.entity.attribute.AttributeDefinition#DECIMAL_SEPARATOR}<br>
  * <li>{@link is.codion.framework.domain.entity.attribute.AttributeDefinition#USE_LEXICAL_STRING_COMPARATOR}<br>